+++ sql/atxs/atxs.go
@@ -27,7 +27,8 @@ func Get(db sql.Executor, id types.ATXID) (atx *types.VerifiedActivationTx, err 
 
 		baseTickHeight := uint64(stmt.ColumnInt64(1))
 		tickCount := uint64(stmt.ColumnInt64(2))
-		atx, err = v.Verify(baseTickHeight, tickCount), nil
+		atx = v.Verify(baseTickHeight, tickCount)
+		err = nil
 		return true
 	}
